misc/HtmlTestSuite/results/test-0662.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>grid items and abspos flex children</title>
<link href="mailto:dgrogan@chromium.org" rel="author" title="David Grogan"/>
<link href="https://drafts.csswg.org/css-flexbox/#abspos-items" rel="help"/>
<link href="https://drafts.csswg.org/css-grid/#algo-overview" rel="help" title="Note at bottom of this section: 'the size of a grid item which is stretched is also considered definite.'"/>
<link href="https://crbug.com/1091588" rel="bookmark"/>
<link href="https://crbug.com/1018439" rel="bookmark"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

<style>
.grid {
  display: grid;
}

.flexbox {
  display: flex;
}

.relpos {
  position: relative;
}

.abspos {
  position: absolute;
  border: 1px solid;
}

/* These are just for making the test look better. */
p {
  margin: 0px;
}
p + div {
  margin-bottom: 40px;
}
</style>

</head><body><p>This crashed Chrome 84.</p>
<div class="grid">
  <div class="relpos flexbox" data-expected-height="20">
    <div class="abspos">Test</div>
    <div style="height: 20px;"/>
  </div>
</div>

<p>Outer is a regular div.</p>
<div>
  <div class="relpos flexbox" data-expected-height="20">
    <div class="abspos">Test</div>
    <div style="height: 20px;"/>
  </div>
</div>

<p>relpos element is a regular div, not flexbox.</p>
<div class="grid">
  <div class="relpos" data-expected-height="20">
    <div class="abspos">Test</div>
    <div style="height: 20px;"/>
  </div>
</div>

<p>Relpos flexbox has a % height descendant. The height of the flexbox grid item is supposed to be definite, allowing the %height flex item to resolve.

This fails in chrome because of https://crbug.com/1018439</p>
<div class="grid">
  <div class="relpos flexbox" data-expected-height="20">
    <div class="abspos">Test</div>
    <div style="height: 20px;"/>
    <div data-expected-height="10" style="height: 50%"/>
  </div>
</div>

<script>
  checkLayout('.relpos');
  // Changing the abspos border shouldn't change the size of any inflow element.
  for (abspos of document.querySelectorAll(".abspos")) {
    abspos.style.borderWidth = "10px";
  }
  checkLayout('.relpos');
</script>
</body></html>
